refactor: BonusNumber를 LottoNumber로 변경
