+++ src/main/java/lotto/model/LottoMatcher.java
@@ -5,12 +5,12 @@ import java.util.List;
 public class LottoMatcher {
     public static final String ERROR_DUPLICATION_BONUS_NUMBER = "보너스 볼 번호가 당첨 번호와 중복입니다.";
     private final WinningNumbers winningNumbers;
-    private final BonusNumber bonusNumber;
+    private final LottoNumber bonusNumber;
 
     public LottoMatcher(List<Integer> winningNumbers, Integer bonusNumber) {
         validateDuplicateBonusNumber(winningNumbers, bonusNumber);
         this.winningNumbers = new WinningNumbers(winningNumbers);
-        this.bonusNumber = new BonusNumber(bonusNumber);
+        this.bonusNumber = new LottoNumber(bonusNumber);
     }
 
     public LottoResult getWinningResult(Lottos lottos) {
